third_party/blink/web_tests/fast/events/constructors/mouse-event-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/input-modifiers.js"></script>
</head>
<body>
<script>

description("This tests the constructor for the MouseEvent DOM class.");

var testObject = {nyannyan: 123};
var testDiv = document.createElement("div");
var xhr = new XMLHttpRequest;

// No initializer is passed.
shouldBe("new MouseEvent('eventType').bubbles", "false");
shouldBe("new MouseEvent('eventType').cancelable", "false");
shouldBe("new MouseEvent('eventType').view", "null");
shouldBe("new MouseEvent('eventType').detail", "0");
shouldBe("new MouseEvent('eventType').screenX", "0");
shouldBe("new MouseEvent('eventType').screenY", "0");
shouldBe("new MouseEvent('eventType').clientX", "0");
shouldBe("new MouseEvent('eventType').clientY", "0");
shouldBe("new MouseEvent('eventType').ctrlKey", "false");
shouldBe("new MouseEvent('eventType').shiftKey", "false");
shouldBe("new MouseEvent('eventType').altKey", "false");
shouldBe("new MouseEvent('eventType').metaKey", "false");
shouldBe("new MouseEvent('eventType').button", "0");
shouldBe("new MouseEvent('eventType').buttons", "0");
shouldBe("new MouseEvent('eventType').relatedTarget", "null");

// bubbles is passed.
shouldBe("new MouseEvent('eventType', { bubbles: false }).bubbles", "false");
shouldBe("new MouseEvent('eventType', { bubbles: true }).bubbles", "true");

// cancelable is passed.
shouldBe("new MouseEvent('eventType', { cancelable: false }).cancelable", "false");
shouldBe("new MouseEvent('eventType', { cancelable: true }).cancelable", "true");

// view is passed.
// Window objects.
shouldBe("new MouseEvent('eventType', { view: window }).view", "window");
shouldBe("new MouseEvent('eventType', { view: this }).view", "this");

// Non-window objects.
shouldThrow("new MouseEvent('eventType', { view: testObject }).view");
shouldThrow("new MouseEvent('eventType', { view: document }).view");
shouldBe("new MouseEvent('eventType', { view: undefined }).view", "null");
shouldBe("new MouseEvent('eventType', { view: null }).view", "null");
shouldThrow("new MouseEvent('eventType', { view: false }).view");
shouldThrow("new MouseEvent('eventType', { view: true }).view");
shouldThrow("new MouseEvent('eventType', { view: '' }).view");
shouldThrow("new MouseEvent('eventType', { view: 'chocolate' }).view");
shouldThrow("new MouseEvent('eventType', { view: 12345 }).view");
shouldThrow("new MouseEvent('eventType', { view: 18446744073709551615 }).view");
shouldThrow("new MouseEvent('eventType', { view: NaN }).view");
// Note that valueOf() is not called, when the left hand side is evaluated.
shouldThrow("new MouseEvent('eventType', { view: {valueOf: function () { return window; } } }).view == window");
shouldBe("new MouseEvent('eventType', { get view() { return window; } }).view", "window");
shouldThrow("new MouseEvent('eventType', { get view() { return 123; } }).view");
shouldThrow("new MouseEvent('eventType', { get view() { throw 'MouseEvent Error'; } })");

// detail is passed.
// Numeric values.
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: 0 }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: 1 }).detail", 1);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: -1 }).detail", -1);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: 2147483647 }).detail", 2147483647);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: -2147483648 }).detail", -2147483648);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: 123.45 }).detail", 123);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: -123.45 }).detail", -123);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: NaN }).detail", 0);
// Non-numeric values.
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: undefined }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: null }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: '' }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: '12345' }).detail", 12345);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: '12345a' }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: 'abc' }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: [] }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: [12345] }).detail", 12345);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: [12345, 67890] }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: {} }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: {moemoe: 12345} }).detail", 0);
shouldBeEqualToNumber("new MouseEvent('eventType', { detail: {valueOf: function () { return 12345; }} }).detail", 12345);

// screenX, screenY, clientX and clientY are passed.
["screenX", "screenY", "clientX", "clientY"].forEach(function (attr) {
    // Numeric values.
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": 0 })." + attr, 0);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": 1 })." + attr, 1);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": -1 })." + attr, -1);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": 2147483647 })." + attr, 2147483647);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": -2147483648 })." + attr, -2147483648);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": 123.45 })." + attr, 123);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": -123.45 })." + attr, -124);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": NaN })." + attr);

    // Non-numeric values.
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": undefined })." + attr, 0);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": null })." + attr, 0);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": '' })." + attr, 0);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": '12345' })." + attr, 12345);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": '12345a' })." + attr);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": 'abc' })." + attr);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": [] })." + attr, 0);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": [12345] })." + attr, 12345);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": [12345, 67890] })." + attr);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": {} })." + attr);
    shouldThrow("new MouseEvent('eventType', { " + attr + ": {moemoe: 12345} })." + attr);
    shouldBeEqualToNumber("new MouseEvent('eventType', { " + attr + ": {valueOf: function () { return 12345; }} })." + attr, 12345);
});

// ctrlKey, altKey, shiftKey and metaKey are passed.
["ctrlKey", "altKey", "shiftKey", "metaKey"].forEach(function (attr) {
    shouldBe("new MouseEvent('eventType', { " + attr + ": false })." + attr, "false");
    shouldBe("new MouseEvent('eventType', { " + attr + ": true })." + attr, "true");
});

// all modifiers are passed correctly.
forEachModifier(function(attr, modifierName) {
    shouldBe("new MouseEvent('eventType', { " + attr + ": false }).getModifierState('" + modifierName + "');", "false");
    shouldBe("new MouseEvent('eventType', { " + attr + ": true }).getModifierState('" + modifierName + "');", "true");
});

// button is passed.
// Numbers within the short range.
shouldBe("new MouseEvent('eventType', { button: 0 }).button", "0");
shouldBe("new MouseEvent('eventType', { button: 1 }).button", "1");
shouldBe("new MouseEvent('eventType', { button: -2 }).button", "-2");
shouldBe("new MouseEvent('eventType', { button: -32768 }).button", "-32768");
shouldBe("new MouseEvent('eventType', { button: 32767 }).button", "32767");

// Numbers that are equal to -1 should be treated as 0.
shouldBe("new MouseEvent('eventType', { button: 65535 }).button", "0");
shouldBe("new MouseEvent('eventType', { button: 9007199254740991 }).button", "0");
shouldBe("new MouseEvent('eventType', { button: -1 }).button", "0");

// Numbers out of the short range.
// 2^{64}-1
shouldBe("new MouseEvent('eventType', { button: 18446744073709551615 }).button", "0");
shouldBe("new MouseEvent('eventType', { button: 12345678901234567890 }).button", "2048");
shouldBe("new MouseEvent('eventType', { button: 123.45 }).button", "123");
shouldBe("new MouseEvent('eventType', { button: NaN }).button", "0");
shouldBe("new MouseEvent('eventType', { button: 65534 }).button", "-2");
shouldBe("new MouseEvent('eventType', { button: -32769 }).button", "32767");
shouldBe("new MouseEvent('eventType', { button: 32768 }).button", "-32768");


// Non-numeric values.
shouldBe("new MouseEvent('eventType', { button: undefined }).button", "0");
shouldBe("new MouseEvent('eventType', { button: null }).button", "0");
shouldBe("new MouseEvent('eventType', { button: '' }).button", "0");
shouldBe("new MouseEvent('eventType', { button: '12345' }).button", "12345");
shouldBe("new MouseEvent('eventType', { button: '12345a' }).button", "0");
shouldBe("new MouseEvent('eventType', { button: 'abc' }).button", "0");
shouldBe("new MouseEvent('eventType', { button: [] }).button", "0");
shouldBe("new MouseEvent('eventType', { button: [12345] }).button", "12345");
shouldBe("new MouseEvent('eventType', { button: [12345, 67890] }).button", "0");
shouldBe("new MouseEvent('eventType', { button: {} }).button", "0");
shouldBe("new MouseEvent('eventType', { button: {moemoe: 12345} }).button", "0");
shouldBe("new MouseEvent('eventType', { button: {valueOf: function () { return 12345; }} }).button", "12345");

// buttons is passed.
// Numbers within the unsigned short range.
shouldBe("new MouseEvent('eventType', { buttons: 0 }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: 1 }).buttons", "1");
shouldBe("new MouseEvent('eventType', { buttons: 65534 }).buttons", "65534");

// Numbers out of the unsigned short range.
// 2^{64}-1
shouldBe("new MouseEvent('eventType', { buttons: 65535 }).buttons", "65535");
shouldBe("new MouseEvent('eventType', { buttons: 9007199254740991 }).buttons", "65535");
shouldBe("new MouseEvent('eventType', { buttons: -1 }).buttons", "65535");
shouldBe("new MouseEvent('eventType', { buttons: 18446744073709551615 }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: 12345678901234567890 }).buttons", "2048");
shouldBe("new MouseEvent('eventType', { buttons: 123.45 }).buttons", "123");
shouldBe("new MouseEvent('eventType', { buttons: NaN }).buttons", "0");

// Non-numeric values.
shouldBe("new MouseEvent('eventType', { buttons: undefined }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: null }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: '' }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: '12345' }).buttons", "12345");
shouldBe("new MouseEvent('eventType', { buttons: '12345a' }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: 'abc' }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: [] }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: [12345] }).buttons", "12345");
shouldBe("new MouseEvent('eventType', { buttons: [12345, 67890] }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: {} }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: {moemoe: 12345} }).buttons", "0");
shouldBe("new MouseEvent('eventType', { buttons: {valueOf: function () { return 12345; }} }).buttons", "12345");

// relatedTarget is passed.
// Valid objects.
shouldBe("new MouseEvent('eventType', { relatedTarget: testDiv }).relatedTarget", "testDiv");
shouldBe("new MouseEvent('eventType', { relatedTarget: document }).relatedTarget", "document");
shouldBe("new MouseEvent('eventType', { relatedTarget: xhr }).relatedTarget", "xhr");
shouldBe("new MouseEvent('eventType', { relatedTarget: window }).relatedTarget", "window");

// Invalid objects.
shouldThrow("new MouseEvent('eventType', { relatedTarget: testObject }).relatedTarget");
shouldBe("new MouseEvent('eventType', { relatedTarget: undefined }).relatedTarget", "null");
shouldBe("new MouseEvent('eventType', { relatedTarget: null }).relatedTarget", "null");
shouldThrow("new MouseEvent('eventType', { relatedTarget: false }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: true }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: '' }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: 'chocolate' }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: 12345 }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: 18446744073709551615 }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { relatedTarget: NaN }).relatedTarget");
// Note that valueOf() is not called, when the left hand side is evaluated.
shouldThrow("new MouseEvent('eventType', { relatedTarget: {valueOf: function () { return testDiv; } } }).relatedTarget == testDiv");
shouldBe("new MouseEvent('eventType', { get relatedTarget() { return testDiv; } }).relatedTarget", "testDiv");
shouldThrow("new MouseEvent('eventType', { get relatedTarget() { return 123; } }).relatedTarget");
shouldThrow("new MouseEvent('eventType', { get relatedTarget() { throw 'MouseEvent Error'; } })");

// All initializers are passed.
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).bubbles", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).cancelable", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).view", "window");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).detail", "111");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).screenX", "222");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).screenY", "333");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).clientX", "444");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).clientY", "555");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).ctrlKey", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).shiftKey", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).altKey", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).metaKey", "true");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).button", "666");
shouldBe("new MouseEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, screenX: 222, screenY: 333, clientX: 444, clientY: 555, ctrlKey: true, shiftKey: true, altKey: true, metaKey: true, button: 666, buttons: 777, relatedTarget: testDiv }).relatedTarget", "testDiv");
</script>
</body>
</html>
